+++ winwin-be-api/src/main/java/com/dpm/winwin/api/member/application/MemberQueryService.java
@@ -1,7 +1,7 @@
 package com.dpm.winwin.api.member.application;
 
 import com.dpm.winwin.api.common.error.exception.custom.BusinessException;
-import com.dpm.winwin.domain.repository.member.dto.MemberDetailResponse;
+import com.dpm.winwin.domain.repository.member.dto.response.MemberDetailResponse;
 import com.dpm.winwin.domain.repository.member.impl.CustomMemberRepositoryImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
